+++ src/main/java/com/prgrms/jpa/controller/dto/post/response/CreatePostResponse.java
@@ -1,0 +1,11 @@
+package com.prgrms.jpa.controller.dto.post.response;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class CreatePostResponse {
+
+    private final Long id;
+}
