+++ src/main/java/com/prgrms/jpa/controller/dto/user/CreateUserRequest.java
@@ -2,6 +2,7 @@ package com.prgrms.jpa.controller.dto.user;
 
 import lombok.Builder;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
@@ -9,6 +10,8 @@ import javax.validation.constraints.Positive;
 import javax.validation.constraints.Size;
 
 @Getter
+@RequiredArgsConstructor
+@Builder
 public class CreateUserRequest {
 
     @NotBlank(message = "이름은 필수 입력사항입니다.")
@@ -22,11 +25,4 @@ public class CreateUserRequest {
     @NotBlank(message = "취미는 필수 입력사항입니다.")
     @Size(max = 30, message = "취미는 {max}글자 이하로 입력할 수 있습니다.")
     private final String hobby;
-
-    @Builder
-    public CreateUserRequest(String name, int age, String hobby) {
-        this.name = name;
-        this.age = age;
-        this.hobby = hobby;
-    }
 }
