+++ src/main/java/com/prgrms/jpa/controller/dto/user/CreateUserResponse.java
@@ -1,13 +1,11 @@
 package com.prgrms.jpa.controller.dto.user;
 
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
 @Getter
+@RequiredArgsConstructor
 public class CreateUserResponse {
 
     private final long id;
-
-    public CreateUserResponse(long id) {
-        this.id = id;
-    }
 }
